+++ test/libsolidity/ASTJSON/enum_value_definition_natspec.sol
@@ -1,0 +1,10 @@
+enum Color {
+	Red,
+	/// @notice example of notice
+	/// @dev example of dev
+	Green,
+	/// @dev example of dev
+	Blue
+}
+
+// ----