+++ test/libsolidity/syntaxTests/natspec/docstring_enum_value_definition.sol
@@ -1,0 +1,11 @@
+contract C {
+	enum Color {
+		Red,
+		/// @notice example of notice
+		/// @dev example of dev
+		Green,
+		/// @dev example of dev
+		Blue
+	}
+}
+// ----